fix(review): extract _has_dynamic_skill_name to _skill_helpers to resolve cross-import
rules/CLAUDE.md prohibits cross-imports between rule modules.
Moved _has_dynamic_skill_name (and its _SKILL_TOKEN_RE) from rules_skills.py
to recipe/_skill_helpers.py so both rules_skills.py and rules_backend_compat.py
can import it without violating the rule.
